tools/code_summarize/requirements.txt
openai
itsprompt
pathspec
python-dotenvpip install 